% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_contingency.R
\name{batch_contingency}
\alias{batch_contingency}
\title{construct a table with contingency values comparing many configuration of type (a vs. b)}
\usage{
batch_contingency(a, b, universe_size, sets = NULL)
}
\arguments{
\item{a}{sets a. if of type list, each component is interpreted as a set.
If a simple vector of characters or integers, interpreted as indexes into
the list of sets. It is advantageous when many a sets are the same.}

\item{b}{sets b to match a;
if type list, each component is interpreted as a set;
if a simple vector, each component is interpreted as an index into 'sets'}

\item{universe_size}{integer, the total number of elements that sets a, b were drawn from}

\item{sets}{list with vectors, used when 'a' or 'b' are not lists}
}
\value{
data table with one row per item, column id and a,b,c,d
}
\description{
construct a table with contingency values comparing many configuration of type (a vs. b)
}
