% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_fisher.R
\name{batch_fisher}
\alias{batch_fisher}
\title{compute fisher comparisons of paired}
\usage{
batch_fisher(data, precomputed = NULL)
}
\arguments{
\item{data}{data.table with contingency values as output by batchContingency}

\item{precomputed}{data table with precomputed p-values and odds-ratios}
}
\value{
data table with contingency data and fisher output for each item in data
}
\description{
compute fisher comparisons of paired
}
