% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_contingency.R
\name{contingency_triple}
\alias{contingency_triple}
\title{summarize overlap of two sets into a vector (partial contingency vector)}
\usage{
contingency_triple(a, b)
}
\arguments{
\item{a}{vector, a set of elements, can be 'numeric' or 'character'}

\item{b}{vector, a set of elements, can be 'numeric' or 'character'}
}
\value{
vector with three components (count_11, count_10, count_01)
}
\description{
summarize overlap of two sets into a vector (partial contingency vector)
}
\keyword{internal}
