% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_contingency.R
\name{contingency_vector}
\alias{contingency_vector}
\title{summarize overlap of two sets into a vector that is ready for a contingency matrix}
\usage{
contingency_vector(a, b, universe_size)
}
\arguments{
\item{a}{vector, a set of elements, can be 'numeric' or 'character'}

\item{b}{vector, a set of elements, can be 'numeric' or 'character'}

\item{universe_size}{integer, total number of elements}
}
\value{
vector with four components (count_11, count_10, count_01, count_00)
}
\description{
summarize overlap of two sets into a vector that is ready for a contingency matrix
}
