% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_fisher.R
\name{fisher2x2}
\alias{fisher2x2}
\title{compute fisher p-values and odds-ratios from a table of contingency values}
\usage{
fisher2x2(x)
}
\arguments{
\item{x}{data table with column $a, $b, $c $d}
}
\value{
an augmented data table with $p.value and $odds.ratio
}
\description{
compute fisher p-values and odds-ratios from a table of contingency values
}
\keyword{internal}
