% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute.R
\name{precompute_fisher2x2}
\alias{precompute_fisher2x2}
\title{precompute a data table with fisher p-values and odds ratios}
\usage{
precompute_fisher2x2(universe_size, vals = seq(0, 20), rotate = FALSE)
}
\arguments{
\item{universe_size}{integer, large number}

\item{vals}{vector of integer, values to consider in first three cells in the confusion matrix}

\item{rotate}{logical, set TRUE to obtain full table of combinations, leave FALSE
to consider column "d" as the special column that holds constrained counts}
}
\value{
data.table with columns (a,b,c,d) that describe entries in a contingency
table and columns (p.value, odds.ratio) that capture output from fisher.test
}
\description{
precompute a data table with fisher p-values and odds ratios
}
