% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute.R
\name{precompute_fisher_contingency}
\alias{precompute_fisher_contingency}
\title{precompute fisher p-values and odds-ratios from a table of contingency values}
\usage{
precompute_fisher_contingency(x)
}
\arguments{
\item{x}{data table with count columns (count_11, count_10, count_01, count_00)}
}
\value{
a table with the contingency values, and new columns with p.value and odds.ratio
}
\description{
precompute fisher p-values and odds-ratios from a table of contingency values
}
