% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute.R
\name{rotate_abcd}
\alias{rotate_abcd}
\title{produce a new table with columns rotated.}
\usage{
rotate_abcd(x)
}
\arguments{
\item{x}{data table with columns a,b,c,d}
}
\value{
data table with same structure as x, but columns a,b,c,d rotated
}
\description{
This creates configurations that are expected to have the same properties under
fisher.test.
}
