% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute.R
\name{precompute_fisher}
\alias{precompute_fisher}
\title{precompute fisher p-values and odds ratios}
\usage{
precompute_fisher(universe_size, a_sizes = seq(0, universe_size),
  b_sizes = a_sizes, max_or = NULL)
}
\arguments{
\item{universe_size}{integer, large number}

\item{a_sizes}{vector of integers, values to consider for sizes of set a}

\item{b_sizes}{vector of integers, values to consider for szies of set b}

\item{max_or}{numeric, maximal population odds-ratio;
When max_or is NULL, the precomputed configurations contain all scenarios
that have sets a and b compatible with the previous arguments.
When max_or is a number, some configurations are omitted if they are unlikely.}
}
\value{
data.table with columns (a,b,c,d) that describe entries in a contingency
table and columns (p.value, odds.ratio) that capture output from fisher.test
}
\description{
Imagine that two sets a and b are composed of items from some superset.
These sets are compared, which leads to a contingency table with components
count_11, count_10, count_01, and count_00. The contingency table can be
processed using a fisher test to obtain a p-value and an odds-ratio.
}
